requirements.txt
tensorflow~=2.3.0
torch~=1.7.1
numpy~=1.18.5
torchvision~=0.8.2
pillow~=8.1.0
sklearn~=0.0
scikit-learn~=0.24.1
tqdm~=4.56.2
matplotlib~=3.3.2